feat: mnemonic attentions
