+++ assets/locales/ar/ar.json
@@ -3,7 +3,7 @@
   "StakingInfoUnDelegationDay": " ({{day}} يوم متبقي؟)",
   "StakingInfoUnDelegationDays": " ({{days}} أيام متبقية)",
   "Welcome_haqq_pin": "مرحبًا بك في HAQQ Wallet",
-  "_hash": "d0b5f36e6ac98d2d7ed6835c13ab490bf8837ba95c47d6faa17396fbf2f26d46",
+  "_hash": "8a59bd346d2f213177530714b71fafcd77f1fa24950b1c9e8c9adfa7104dc934",
   "accept": "قبول",
   "accountInfoNftTabTitle": "العملات غير القابلة للاستبدال",
   "accountInfoTokensTabTitle": "العملات",
@@ -18,7 +18,7 @@
   "amountISLM": "{{amount}} ISLM",
   "apply": "تطبيق",
   "average": "المتوسط",
-  "backupCreateRecoveryAgreement": "أدرك أنه إذا فقدت عبارة الاسترداد الخاصة بي، فلن أتمكن من استعادة الوصول إلى حسابي",
+  "backupCreateRecoveryAgreement": "العبارة السرية هي مفتاحي الخاص ولن أشاركها مع أي شخص",
   "backupCreateRecoveryPhrase": "عبارة الاسترداد الخاصة بك",
   "backupCreateRecoverySaveWords": "اكتب هذه الكلمات أو انسخها بالترتيب الصحيح واحفظها في مكان آمن.",
   "backupCreateRecoveryWarningMessage": "إذا فقدت عبارة الاسترداد الخاصة بك، لن تتمكن من الوصول إلى أموالك، لأنه لن يتمكن أحد من استعادتها.",
