+++ assets/locales/en/en.json
@@ -3,7 +3,7 @@
   "StakingInfoUnDelegationDay": " ({{day}} day left)",
   "StakingInfoUnDelegationDays": " ({{days}} days left)",
   "Welcome_haqq_pin": "Welcome to HAQQ Wallet",
-  "_hash": "ab34a690eb660dea9d829a7b1b9cdbbbbb266735341dd0b6f3cda4560d54f926",
+  "_hash": "11e244ed30a5ce81abd18764e5daf82433d9f3c694c4823ca7a8258cefe2e936",
   "accept": "Accept",
   "accountInfoNftTabTitle": "NFTs",
   "accountInfoTokensTabTitle": "Tokens",
@@ -12,6 +12,7 @@
   "actionSheetKeepEditing": "Keep Editing",
   "actionSheetMessage": "Are you sure you want to discard your changes? They will be lost.",
   "address": "Address",
+  "addressNetwork": "Address network",
   "adjustment": "Adjustment",
   "amount": "{{amount}} {{symbol}}",
   "amountAISLM": "{{amount}} {{symbol}}",
@@ -20,7 +21,7 @@
   "average": "Average",
   "backendPushTransferDescription": "From {{address_start}}...{{address_end}}",
   "backendPushTransferTitle": "Received {{amount}} {{token}}",
-  "backupCreateRecoveryAgreement": "I understand that losing my recovery phrase means that I won't be able to recover my account.",
+  "backupCreateRecoveryAgreement": "The secret phrase is my private key, and I will not share it with anyone.",
   "backupCreateRecoveryPhrase": "Your recovery phrase",
   "backupCreateRecoverySaveWords": "Write down or copy these words in the right order and save them somewhere safe.",
   "backupCreateRecoveryWarningMessage": "If you lose your recovery phrase, you will lose access to your funds, and no one can help recover it.",
@@ -568,6 +569,7 @@
   "rewardCreatingFirstAccount": "Reward for creating your first account.",
   "save": "Save",
   "selectAccount": "Select Account",
+  "selectCryptoToSend": "Select crypto to send",
   "sent": "Sent",
   "settingsAboutDiscord": "Discord",
   "settingsAboutDocuments": "Legal Documents",
@@ -993,9 +995,11 @@
   "turnOnPushNotificationsDescription": "We can notify you when something important happens, such as: The transaction was successfully sent. Funds have arrived in the account. We detect any risks for the security of your account",
   "turnOnPushNotificationsSkip": "Not now",
   "turnOnPushNotificationsTitle": "Turn on push notifications",
+  "typeAssets": "Type assets",
   "unknownError": "Unknown error",
   "unknownQrCode": "Invalid QR code. Please retry.",
   "unrecognizedDataFormat": "Unrecognized data format. Please try again.",
+  "unsupportedNetworkError": "The network associated with the address you entered is not supported. Please enter a different address",
   "validatorInfoHide": "Hide",
   "validatorInfoReward": "Reward",
   "validatorInfoShowOther": "Show other",
