+++ assets/locales/tr/tr.json
@@ -3,7 +3,7 @@
   "StakingInfoUnDelegationDay": "({{day}} gün kaldı)",
   "StakingInfoUnDelegationDays": "({{days}} gün kaldı)",
   "Welcome_haqq_pin": "HAQQ Cüzdanına Hoş Geldiniz",
-  "_hash": "d0c8e4ad29ad39e8a9e48fe265c718ccf48ac0f862072c8fd2ac14ac929d5dcd",
+  "_hash": "a6e765bf836b67be2ecde295db165b7b2ef184e7899c33e17788dcf3b1db62c9",
   "accept": "Kabul Et",
   "accountInfoNftTabTitle": "NFT’ler",
   "accountInfoTokensTabTitle": "Tokenlar",
@@ -18,7 +18,7 @@
   "amountISLM": "{{amount}} {{symbol}}",
   "apply": "Uygula",
   "average": "Ortalama",
-  "backupCreateRecoveryAgreement": "Kurtarma cümlemi kaybedersem, hesabıma erişimi geri yükleyemeyeceğimi anlıyorum",
+  "backupCreateRecoveryAgreement": "Gizli ifade benim özel anahtarımdır ve onu kimseyle paylaşmayacağım.",
   "backupCreateRecoveryPhrase": "Kurtarma cümleniz",
   "backupCreateRecoverySaveWords": "Bu kelimeleri doğru sırayla yazın veya kopyalayın ve güvenli bir yerde saklayın.",
   "backupCreateRecoveryWarningMessage": "Kurtarma cümlenizi kaybederseniz, fonlarınıza erişemezsiniz, çünkü kimse bunu geri yükleyemez.",
